+++ powershell/windows-install/bullet.ps1
@@ -24,12 +24,14 @@ try {
     # Configure/compile
     cmake .. -G Ninja -DCMAKE_BUILD_TYPE="$BuildType" -DBUILD_SHARED_LIBS=OFF -DUSE_MSVC_RUNTIME_LIBRARY_DLL=ON -DBUILD_BULLET3=OFF -DBUILD_BULLET2_DEMOS=OFF -DBUILD_EXTRAS=OFF -DBUILD_UNIT_TESTS=OFF -DBUILD_PYBULLET=OFF -DINSTALL_LIBS=ON -DCMAKE_INSTALL_PREFIX="C:\bullet"
     ninja
+    if($LastExitCode -ne 0) { throw }
 
     # Remove the older install (if it exists)
     Remove-Item -Recurse -Force -ErrorAction SilentlyContinue -Path C:\bullet
 
     ## Install
     ninja install
+    if($LastExitCode -ne 0) { throw }
     
     # Delete our working directory
     cd $invocationDir
